Separate MapComponentHandler from map-registration and add update map handler.
